requirements.txt
pytesseract==0.3.10
Pillow==10.1.0
pytest==7.4.3
pdf2image==1.16.3
